Chrome opens in app mode with contained profile. Auto-close processes.
Previously, if chrome was already open, nodebug would fail because it
wouldn't be able to get a lock on the default profile. Now a profile is
saved in this local folder (and is ignored in git/npm) and chrome starts
in "app mode" e.g. no browser chrome (back buttons, address bar, etc).
This makes for a more focused experience.

Processes also close automatically, to avoid having to manually
close/restart the debugging and chrome, unless the `--keep` option is
passed.